use a dependency injection pattern to avoid loading the http server stuff when running in the browser
