+++ browser.js
@@ -1,0 +1,2 @@
+
+module.exports = require('./inject')(function () {})
